refactor: move into package
