+++ tests/affine-local/e2e/all-page.spec.ts
@@ -1,4 +1,4 @@
-import { test } from '@affine-test/kit/playwright';
+import { test } from '@affine-test/kit/lib/playwright';
 import {
   changeFilter,
   checkDatePicker,
