+++ tests/affine-local/e2e/blocksuite/editor.spec.ts
@@ -1,4 +1,4 @@
-import { test } from '@affine-test/kit/playwright';
+import { test } from '@affine-test/kit/lib/playwright';
 import { openHomePage } from '@affine-test/kit/utils/load-page';
 import { newPage, waitEditorLoad } from '@affine-test/kit/utils/page-logic';
 import type { Page } from '@playwright/test';
