+++ tests/affine-local/e2e/debug-init-page.spec.ts
@@ -5,6 +5,7 @@ test('should have page0', async ({ page }) => {
   await page.goto('http://localhost:8080/_debug/init-page.html');
   await page.waitForSelector('v-line');
   const pageId = await page.evaluate(async () => {
+    // @ts-expect-error
     return globalThis.page.id;
   });
   expect(pageId).toBe('page0');
