+++ tests/affine-local/e2e/drag-page-to-trash-folder.spec.ts
