+++ tests/affine-local/e2e/exception-page.spec.ts
