+++ tests/affine-local/e2e/image-preview.spec.ts
