+++ tests/affine-local/e2e/invite-code-page.spec.ts
