+++ tests/affine-local/e2e/local-first-avatar.spec.ts
