+++ tests/affine-local/e2e/local-first-collections-items.spec.ts
