+++ tests/affine-local/e2e/local-first-delete-page.spec.ts
