+++ tests/affine-local/e2e/local-first-delete-workspace.spec.ts
