+++ tests/affine-local/e2e/local-first-favorite-page.spec.ts
