+++ tests/affine-local/e2e/local-first-favorites-items.spec.ts
