+++ tests/affine-local/e2e/local-first-new-page.spec.ts
