+++ tests/affine-local/e2e/local-first-openpage-newtab.spec.ts
