+++ tests/affine-local/e2e/local-first-show-delete-modal.spec.ts
