+++ tests/affine-local/e2e/local-first-trash-page.spec.ts
