+++ tests/affine-local/e2e/local-first-workspace.spec.ts
