+++ tests/affine-local/e2e/open-affine.spec.ts
@@ -1,4 +1,4 @@
-import { test } from '@affine-test/kit/playwright';
+import { test } from '@affine-test/kit/lib/playwright';
 import { openHomePage } from '@affine-test/kit/utils/load-page';
 import { waitEditorLoad } from '@affine-test/kit/utils/page-logic';
 import { createWorkspace } from '@affine-test/kit/utils/workspace';
