+++ tests/affine-local/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@affine-test/affine-local",
+  "private": true,
+  "scripts": {
+    "e2e": "yarn playwright test"
+  },
+  "devDependencies": {
+    "@affine-test/fixtures": "workspace:*",
+    "@affine-test/kit": "workspace:*",
+    "@playwright/test": "^1.35.1"
+  }
+}
