+++ tests/kit/tsconfig.json
@@ -5,5 +5,5 @@
     "noEmit": false,
     "outDir": "lib"
   },
-  "include": ["./*.ts"]
+  "include": ["./*.ts", "utils"]
 }
