+++ tests/unit/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "baseUrl": "..",
+    "noEmit": true
+  },
+  "include": ["**.spec.ts"]
+}
